move enum to its own file
