+++ src/service/ampdoc-impl.js
@@ -59,15 +59,6 @@ const AmpDocSignals = {
 };
 
 /**
- * Registred singleton on AMP doc.
- * @enum {number}
- */
-export const AMPDOC_SINGLETON_NAME = {
-  TRACKING_IFRAME: 1,
-  LINKER: 2,
-};
-
-/**
  * This service helps locate an ampdoc (`AmpDoc` instance) for any node,
  * either in the single-doc or shadow-doc environments.
  *
@@ -317,7 +308,7 @@ export class AmpDoc {
     /** @public @const {!Window} */
     this.win = win;
 
-    /** @private {!Object<AMPDOC_SINGLETON_NAME, boolean>} */
+    /** @private {!Object<../enums.AMPDOC_SINGLETON_NAME, boolean>} */
     this.registeredSingleton_ = map();
 
     /** @public @const {?AmpDoc} */
@@ -765,7 +756,7 @@ export class AmpDoc {
   /**
    * Attempt to register a singleton for each ampdoc.
    * Caller need to handle user error when registration returns false.
-   * @param {AMPDOC_SINGLETON_NAME} name
+   * @param {!../enums.AMPDOC_SINGLETON_NAME} name
    * @return {boolean}
    */
   registerSingleton(name) {
